Adds "sidenote" support
Inspired by https://github.com/clayh53/tufte-jekyll/, this adds
support for hidden asides. I originally wanted to pull in the
complete Tufte style "notes in the margins", but getting it to look
the way I wanted was more of an investment than I was willing to put
in. Instead it's just hidden notes that can be revealed. Effectively
identical to what happens with the Tufte style notes when the page
width is too small to have a margin.