Isolates Co-op terms and adds section hiding
Co-op terms aren't the same as full-time work, so they get their own
section from now on. Also adds section collapsing, which is most
useful for hiding co-op terms by default.

Section hiding is implemented in javascript, such that when the page
loads all the content is displayed, to ensure support for browsers
that do not have, or chose not to have javascript enabled. This uses
a click handler on both the header, and button. The buttons provide a
good visual marker for the user to expand collapse the sections.